alloy_sol_macro_expander
Module expand

Functions which generate Rust code from the Solidity AST.

Structs

ExpCtxt

The expansion context.

ExternCrates

Configurable extern crate dependencies.

NamespacedMap

Mapping namespace -> ident -> T

Functions

anon_name

Returns the name of a parameter, or a generated name if it is None.

expand

The sol! expansion implementation.

expand_type

Expands a Rust type from a Solidity type.

generate_name

Generates an anonymous name from an integer. Used in anon_name.